Rename and move test services
