+++ packages/core/src/openapi/openapi-request-builder.ts
@@ -8,7 +8,7 @@ import type {
 
 /**
  * @experimental This API is experimental and might change in newer versions. Use with caution.
- * Generic request builder for building and executing REST request requests based on apis generated by the OpenAPI generator.
+ * Generic request builder for building and executing requests based on apis generated by the OpenAPI generator.
  * @typeparam ApiT Interface of the generated api. All properties of the interface will be treated as functions.
  * @typeparam FnT Literal string type of the function name, must be a key of `ApiT` and represent a function.
  */
