+++ scripts/generate-test-services.ts
@@ -137,8 +137,8 @@ if (arg === 'e2e' || arg === 'all') {
 
 if (arg === 'openapi' || arg === 'rest' || arg === 'all') {
   generateOpenApi({
-    inputDir: path.resolve('test-resources', 'rest-service-specs'),
-    outputDir: path.resolve('test-packages', 'test-services', 'rest'),
+    inputDir: path.resolve('test-resources', 'openapi-service-specs'),
+    outputDir: path.resolve('test-packages', 'test-services', 'openapi'),
     clearOutputDir: true
   });
 }
