+++ test-packages/test-services/openapi/test-service/openapi/.npmignore
