+++ test-packages/test-services/openapi/test-service/openapi/.openapi-generator/VERSION
