+++ test-packages/test-services/openapi/test-service/openapi/model/simple-test-entity.ts
